Removing user property from git clone resources
The vcsrepo plugin attempts to use su, which will only run from a tty.
This prevents using the puppet script from a login script (making it
pretty useless).

For now we're assuming this manifest is run as the intended user anyway
since we're relying on the $id variable, so no harm in just dropping the
user argument.